chore: convert fixture error to warning
