+++ docs/credentials/AgInspectionReport.json
@@ -136,8 +136,8 @@
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:44.662Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..a28KQX59W-X_KeGx4Ns5iZYI7lXHJVeCrqIouO16az_qbYxoGmILRiUTFp_bJt0PcWnaooBCc0T-0hCPt0plAg",
+    "created": "2021-02-16T22:37:10.922Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..OT6Oxb1QmlOfuPflgi-z3oy0g6cIk9iPzctJ_vnz7wB9KZzACbcT9y8bcPt6A8HDO0nw4GM7eFaueLPCOn4jCw",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
