+++ docs/credentials/EcommerceAdditionalProductCodeRegistrationCredential.json
@@ -14,15 +14,15 @@
   "credentialSubject": {
     "id": "did:web:www.acme.com",
     "type": "EcommerceAdditionalProductCodeRegistrationCredential",
-    "productVCid": "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58597a",
-    "addProductCode": 26496222089044,
-    "addProductCodeType": "HTS",
-    "certificateName": "ACME Ecommerce HTS Code Product Registration Certificate"
+    "productVCid": "https://vc.example.com/?queryID=AB06f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58597a",
+    "addProductCode": 35959660352673,
+    "addProductCodeType": "CO",
+    "certificateName": "ACME Ecommerce Country of Origin Code Product Registration Certificate"
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:47.048Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..RRnE65J5B5VQprCXr0d7GUbsYGiUWVJ8TyJCXcajwnvlH0Cydt6GCT0RLSZd9JafsiaeD4aiJQh_EdW02NVHBA",
+    "created": "2021-02-16T22:37:13.281Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..nQ3i7o7uwSTDfEq6ixRKMtKXLuPctQAZj0DKSJMgP7Sx7lHJHveMGG7-y9XE8xhI__GeJB5Hv0znkac9vh9hDw",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
