+++ docs/credentials/EcommerceInvoiceRegistrationCredential.json
@@ -14,19 +14,19 @@
   "credentialSubject": {
     "id": "did:web:www.acme.com",
     "type": "EcommerceInvoiceRegistrationCredential",
-    "invoiceID": "Invoice#517",
+    "invoiceID": "Invoice#99",
     "orderID": [
       "https://vc.example.com/?queryID=AB06f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58597a"
     ],
     "productInOrder": [
-      "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a"
+      "https://vc.example.com/?queryID=6206f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a"
     ],
     "certificateName": "ACME Ecommerce Invoice Registration Credential"
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:46.668Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..Ai_GufpMwnfQD8_LE_US2HlXe1N6POCLw_3FCI3a_uSKw5qM57M5zQYdnzVGmWGqDgFHvms4YBfnVQrG6KRTCg",
+    "created": "2021-02-16T22:37:13.474Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..c0vc0hT56trma5OCcSZiKEfNgPapoaO6HvCG1-JvPsFtidiTRjqIRNSOv_SJRDXMnE1J4VQBtmC6zULze_i6CA",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
