+++ docs/credentials/EcommerceOrderRegistrationCredential.json
@@ -14,17 +14,16 @@
   "credentialSubject": {
     "id": "did:web:www.acme.com",
     "type": "EcommerceOrderRegistrationCredential",
-    "orderID": "Order#692",
+    "orderID": "Order#249",
     "productInOrder": [
-      "https://vc.example.com/?queryID=6206f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a",
       "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a"
     ],
     "certificateName": "ACME Ecommerce Order Registration Credential"
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:46.666Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..VTgRTRtHkvUc2sCCOR8riuOkzK0cATSVlqnhEng4dV4OOyU-4Bnnq8ddoW_ZMNAfoDpQk5Ry2y0ClIZsqFHwCQ",
+    "created": "2021-02-16T22:37:13.511Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..fZWuw12kF2ye1vsqNJxfggHrdj9kDTTYA3Hn3ESD67v4KOAB5-yxSYZwLEc8wY1JqcbdV5w15B-tZyWcsftoAA",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
