+++ docs/credentials/EcommercePackageRegistrationCredential.json
@@ -26,16 +26,14 @@
   "credentialSubject": {
     "id": "did:web:www.fedex.com",
     "type": "EcommercePackageRegistrationCredential",
-    "trackingID": "849804346193",
+    "trackingID": "874435044964",
     "packageItems": [
       {
-        "productReceiptID": "https://example.VC_url.com/?queryID=0xA6D5FeB0DfFfA1dBDebf52aC8b703b8D09d4Eb37DA94C6Cd89EA5765BCDCd1A9",
-        "packingListID": "https://example.VC_url.com/?queryID=0xa2e0efd7cb98391A4d1f4Edb5BcEcbe8f8Cbc9ca7224cc3462b420099ef3051B",
+        "productReceiptID": "https://example.VC_url.com/?queryID=0xBc6B0b13CB07adfAD7F1aFDF34fBaA1a2515ee6A282AfE0cac0c3EcA096d145d",
+        "packingListID": "https://example.VC_url.com/?queryID=0xC7adB4bf639ddB93aec20e4ea7DDB4Fae0aa4EcbBdCB4eeEeBDEaEE3ECfFdAdD",
         "orderNumber": "https://vc.example.com/?queryID=AB06f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58597a",
         "productInOrder": [
-          "https://vc.example.com/?queryID=6206f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a",
-          "https://vc.example.com/?queryID=6206f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a",
-          "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a"
+          "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58599e"
         ]
       }
     ],
@@ -43,8 +41,8 @@
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:46.681Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..lmQVanivnShOMUoEUrX55J50_yt5p6IBM9N3FSm9nSbtlTn3dANUQqVViXO9dqxqA49TYP3iG_qlKFSk0M44AQ",
+    "created": "2021-02-16T22:37:13.495Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..hXGUy_6ohrgPllg99q8b46npKg1huOp7E1nN01cBaMQ2BAGdXcg00Ugg1cQGSMTByU3ZItPzZpNx9Nm2eps7Dg",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
