+++ docs/credentials/EcommercePackingListRegistrationCredential.json
@@ -18,6 +18,9 @@
       {
         "orderNumber": "https://vc.example.com/?queryID=AB06f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58597a",
         "productInOrder": [
+          "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a",
+          "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a",
+          "https://vc.example.com/?queryID=6206f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a",
           "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58599e"
         ]
       }
@@ -26,8 +29,8 @@
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:46.688Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..w3ETWLe6PN_xeK3svhsJb2Dbb8MLIjoHnxFDrPyrY9zJwFVN12zu5LO44SaktS80nUmmt8iFSBOvM8vnJPIaCA",
+    "created": "2021-02-16T22:37:13.474Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..FIZohqbcZSehSX_bFbKEG7w742_sGN1FHUqCtuHyACXBqeLoXy_ldnL4kgvqroqOCQdYWw2fhHdJIbi6pch6CA",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
