+++ docs/credentials/EcommerceProductReceiptRegistrationCredential.json
@@ -14,7 +14,7 @@
   "credentialSubject": {
     "id": "did:web:www.acmecarrier.com",
     "type": "EcommerceProductReceiptRegistrationCredential",
-    "receiptID": "Receipt#294",
+    "receiptID": "Receipt#354",
     "packingListID": [
       "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58597a"
     ],
@@ -22,14 +22,17 @@
       "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58597a"
     ],
     "productInOrder": [
+      "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58599e",
+      "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58599e",
+      "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a",
       "https://vc.example.com/?queryID=7306f1f744a781480c521902a1a1dbf5f1d01e7ea21daf483e7668817e58598a"
     ],
     "certificateName": "ACME Carrier Ecommerce Product Receipt Registration Certificate"
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:46.649Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..DseMeFJV-TKBcozqLamBjF92jbLQmu7KZiII8Zhlz4dNdmFyXXe9KIu9iNOFu_u_afVG7h0TJc0Qh04ZLserCA",
+    "created": "2021-02-16T22:37:13.268Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..3uhbq7L3XMgJMpvZ2QfIYB3m3PAhCWvNvL7ICHetJQDkU2QwmPj46PdJaLNCKaGCS5bPI4rsrwLaCywgunqqBA",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
