+++ docs/credentials/EcommerceProductRegistrationCredential.json
@@ -14,14 +14,14 @@
   "credentialSubject": {
     "id": "did:web:www.acme.com",
     "type": "EcommerceProductRegistrationCredential",
-    "productCode": 7,
+    "productCode": 5,
     "productCodeType": "GTIN",
     "certificateName": "ACME Ecommerce Product Registration Credential"
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:46.645Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..1cBmEJa8PCw9FvpTtD0kJxdU21gn-Bor4lDBcieNG3thhALBRWdh7C6DgjlxdT7DktdzYFwY1jiElJeAYGn3AA",
+    "created": "2021-02-16T22:37:13.243Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..WygOYpyQGEb4XLNnUa-5e_QQhTp4n1AihlHoy8LyTro_emrGdlIRO9ag_qzXVPEsp5P2S7yUr4Lt15LhIhxLCQ",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
