+++ docs/credentials/LegalEntityIdentifierCredential.json
@@ -14,13 +14,13 @@
   "credentialSubject": {
     "id": "did:web:www.acme.com",
     "type": "LegalEntityIdentifierCredential",
-    "leiCode": "RCK7PVDPGYNYUO5S02JQ",
+    "leiCode": "D2UKI5RRLEHXDZPA3CS5",
     "certificateName": "US Legal Entity Certificate"
   },
   "proof": {
     "type": "Ed25519Signature2018",
-    "created": "2021-02-16T22:19:46.639Z",
-    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..GXgPVp_FgSBCOvvlV4jelBrtI7fPy66QpyrkrflHyJkP2XnYHiJ_G_gqG5YIqxV_0a_rRwRqzYGwOD7G8sPPCw",
+    "created": "2021-02-16T22:37:13.255Z",
+    "jws": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..lKpWiJjuExBz3ibwj_6LBZGYk1B4ZCqMdaW5PFUIdPA4oQShdEPU6T7-C5CGIi5o1NUuqwe-bDC7c5O7_irwAQ",
     "proofPurpose": "assertionMethod",
     "verificationMethod": "did:example:123#key-0"
   }
