+++ docs/invalid-fixtures-list.json
@@ -1,0 +1,8 @@
+[
+  "https://w3id.org/traceability/schemas/EcommerceInvoiceRegistrationEvidenceDocument.json",
+  "https://w3id.org/traceability/schemas/EcommerceOrderRegistrationEvidenceDocument.json",
+  "https://w3id.org/traceability/schemas/EcommercePackageRegistrationCredential.json",
+  "https://w3id.org/traceability/schemas/EcommercePackageRegistrationEvidenceDocument.json",
+  "https://w3id.org/traceability/schemas/EcommercePackingListRegistrationEvidenceDocument.json",
+  "https://w3id.org/traceability/schemas/LEIevidenceDocument.json"
+]
